utilize a single audit command
